% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_shape_log_message}
\alias{df_shape_log_message}
\title{Creates a log of dataframe shape, and entry missingness}
\usage{
df_shape_log_message(data_to_log)
}
\arguments{
\item{data_to_log}{\itemize{
\item dataset to create a log message from
}}
}
\value{
a string summarising dataframe shape and missing info
}
\description{
Creates a log of dataframe shape, and entry missingness
}
\examples{
df_shape_log_message(cars)
}
